Deletes some unimportant apis and splits converter api to some class file.
