+++ KeLi.Common.Revit/Builders/GridWallBuilder.cs
@@ -1,11 +1,59 @@
-﻿using Autodesk.Revit.DB;
+﻿/*
+ * MIT License
+ *
+ * Copyright(c) 2019 KeLi
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+/*
+             ,---------------------------------------------------,              ,---------,
+        ,----------------------------------------------------------,          ,"        ,"|
+      ,"                                                         ,"|        ,"        ,"  |
+     +----------------------------------------------------------+  |      ,"        ,"    |
+     |  .----------------------------------------------------.  |  |     +---------+      |
+     |  | C:\>FILE -INFO                                     |  |  |     | -==----'|      |
+     |  |                                                    |  |  |     |         |      |
+     |  |                                                    |  |  |/----|`---=    |      |
+     |  |              Author: KeLi                          |  |  |     |         |      |
+     |  |              Email: kelistudy@163.com              |  |  |     |         |      |
+     |  |              Creation Time: 04/16/2020 01:57:20 PM |  |  |     |         |      |
+     |  | C:\>_                                              |  |  |     | -==----'|      |
+     |  |                                                    |  |  |   ,/|==== ooo |      ;
+     |  |                                                    |  |  |  // |(((( [66]|    ,"
+     |  `----------------------------------------------------'  |," .;'| |((((     |  ,"
+     +----------------------------------------------------------+  ;;  | |         |,"
+        /_)_________________________________________________(_/  //'   | +---------+
+           ___________________________/___  `,
+          /  oooooooooooooooo  .o.  oooo /,   \,"-----------
+         / ==ooooooooooooooo==.o.  ooo= //   ,`\--{)B     ,"
+        /_==__==========__==_ooo__ooo=_/'   /___________,"
+*/
+
+using Autodesk.Revit.DB;
 
 using KeLi.Common.Revit.Filters;
 
 namespace KeLi.Common.Revit.Builders
 {
     /// <summary>
-    /// Grid wall builder.
+    ///     Grid wall builder.
     /// </summary>
     public static class GridWallBuilder
     {
