--- KeLi.Common.Revit/Builders/PipeParameter.cs
@@ -1,90 +1,0 @@
-﻿/*
- * MIT License
- *
- * Copyright(c) 2019 KeLi
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-/*
-             ,---------------------------------------------------,              ,---------,
-        ,----------------------------------------------------------,          ,"        ,"|
-      ,"                                                         ,"|        ,"        ,"  |
-     +----------------------------------------------------------+  |      ,"        ,"    |
-     |  .----------------------------------------------------.  |  |     +---------+      |
-     |  | C:\>FILE -INFO                                     |  |  |     | -==----'|      |
-     |  |                                                    |  |  |     |         |      |
-     |  |                                                    |  |  |/----|`---=    |      |
-     |  |              Author: KeLi                          |  |  |     |         |      |
-     |  |              Email: kelistudy@163.com              |  |  |     |         |      |
-     |  |              Creation Time: 01/22/2020 11:22:30 AM |  |  |     |         |      |
-     |  | C:\>_                                              |  |  |     | -==----'|      |
-     |  |                                                    |  |  |   ,/|==== ooo |      ;
-     |  |                                                    |  |  |  // |(((( [66]|    ,"
-     |  `----------------------------------------------------'  |," .;'| |((((     |  ,"
-     +----------------------------------------------------------+  ;;  | |         |,"
-        /_)_________________________________________________(_/  //'   | +---------+
-           ___________________________/___  `,
-          /  oooooooooooooooo  .o.  oooo /,   \,"-----------
-         / ==ooooooooooooooo==.o.  ooo= //   ,`\--{)B     ,"
-        /_==__==========__==_ooo__ooo=_/'   /___________,"
-*/
-
-using System;
-
-using Autodesk.Revit.DB;
-
-namespace KeLi.Common.Revit.Builders
-{
-    /// <summary>
-    ///     Pipe parameter.
-    /// </summary>
-    public class PipeParameter
-    {
-        /// <summary>
-        ///     Pipe parameter.
-        /// </summary>
-        /// <param name="systemId"></param>
-        /// <param name="typeId"></param>
-        /// <param name="lvlId"></param>
-        public PipeParameter(ElementId systemId, ElementId typeId, ElementId lvlId)
-        {
-            SystemId = systemId ?? throw new ArgumentNullException(nameof(systemId));
-
-            TypeId = typeId ?? throw new ArgumentNullException(nameof(typeId));
-
-            LevelId = lvlId ?? throw new ArgumentNullException(nameof(lvlId));
-        }
-
-        /// <summary>
-        ///     System id.
-        /// </summary>
-        public ElementId SystemId { get; set; }
-
-        /// <summary>
-        ///     Type id.
-        /// </summary>
-        public ElementId TypeId { get; set; }
-
-        /// <summary>
-        ///     Level id.
-        /// </summary>
-        public ElementId LevelId { get; set; }
-    }
-}
